requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
duckdb-engine>=0.9.2
pandas>=2.2.0
pydantic>=2.6.0
python-multipart>=0.0.6
requests>=2.31.0
kagglehub>=0.1.0 